CreateCard Testing added

+++ Models/CardDeckModel.cs
@@ -84,8 +84,7 @@ namespace Models
         {
             List<CardBaseModel> cardsToReturn = new();
 
-            if (amount < 0) return cardsToReturn;
-            if (amount > Cards.Count) return cardsToReturn;
+            if (amount < 0 || amount > Cards.Count) return cardsToReturn;
 
             cardsToReturn = Cards.Take(amount).ToList();
             Cards.RemoveRange(0, amount);
